Using app yaml for deployment on ecs
